Feat : 관광지 거리 계산 메서드 추가

+++ src/main/java/com/ssafy/bridge/bookmark/dao/BookMarkDaoImpl.java
@@ -110,8 +110,8 @@ public class BookMarkDaoImpl implements BookMarkDao {
 						rs.getInt("readcount"),
 						rs.getInt("sido_code"),
 						rs.getInt("gugun_code"),
-						rs.getDouble("latitude"),
-						rs.getDouble("longitude"),
+						rs.getBigDecimal("latitude"),
+						rs.getBigDecimal("longitude"),
 						rs.getString("mlevel"),
 						rs.getString("description"),
 						rs.getString("member_id")
